tests are successful

+++ raw_to_digi/src/SiStripFEDBuffer.cxx
@@ -76,7 +76,7 @@ bool FEDBuffer::isZeroSuppressed() const {
 
 bool FEDBuffer::isValid() const {
   // Extract if commissioning info are valid or not
-  const uint32_t daq1 = static_cast<uint32_t>(feHeader_->daqRegister());
+  const uint32_t daq1 = feHeader_->daqRegister();
   const uint16_t temp = static_cast<uint16_t>((daq1 >> 8) & 0x3);
   return (temp == uint16_t(1));
 }
